Initial commit for fwtiling.py - the new location for FW relating tiling functionality, mainly the FWTiler class to encapsulate tiling settings, wrap tiling functions, and perform tiling operations. Base of class is done, including wrappers for tiling/TaipanTile functions, but existing FW tiling functions are yet to be incorporated.
